i2c: fix in ack sending logic in i2c-microchip
Resolves SAR122362.

"Master reciever mode" in the were not correctly sending ACKs/NACKs
in the interrupt handler.
As a fix for this, the handling of STATUS_M_SLAR_ACK,
STATUS_M_RX_DATA_ACKED and STATUS_M_RX_DATA_NACKED were changed to
bring them in line with the bare metal driver.

Fixes:  6c7955c3bfd ("i2c: microchip: Add driver for Microchip
PolarFire SoC")

Signed-off-by: shravan kumar <shravan.chippa@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>