soc/microchip: reorder operations in generic service
Previously the driver read in a bunch of data and executed a transaction
each and every time a read was done. instead, move the transaction to
the write function and allow only a single read per write. Also use
devm_ version allocations/frees.

This has the side effect of also clearing an kernel oops, probably
caused by the a lack of mallocs.