riscv: dts: microchip: dont clock periphierals on fic clks
"fic" clocks are not suitable for use as a peripheral's clock source
as they are only intended for clock domain crossing to the fabric.

Physically, fixed frequency clocks in the fabric actually provide the
clocks that these peripherals are using, so create those dt entries.
Replace any uses of "ficN" as a clocksource with the suitable fabric
clock.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>