clk: microchip: mpfs: correctly reparent clocks
The mpfs clock driver incorrectly parents the clocks off a 600M
reference clock: the "msspll". This pll is configurable and the real
reference clock is a 100 or 125 MHz oscillator on the board. The msspll
is actually set up by the bootloader and its configuration should be
read by this driver & used as the parent for the cpu/axi/ahb clocks.

Secondly, the rtc clock is based on the external reference, not the
msspll clock as is the case currently. Now that the external reference
is actually present, reparent it too.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>