rtc: mpfs: convert to binary mode
The rtc was using calendar mode with a limited range of what it could
actually count. Calendar mode is largely useless in Linux as the rtc
framework handles the conversions to human readable dates. Therefore
change to binary mode which has a significantly greater range.

The rtc also only was consuming its ahb clock, and not its 1 MHz
reference. Add the reference and calculate the prescaler value from the
rate of the reference rather than getting a value from the device tree.

Also various cleanups requested in the upstreaming process:
- remove oneline mpfs_rtc_set_prescaler function
- dont unconditionally turn off the rtc in the init function
- dont reset the rtc when init is run.
- dont disable the alarm when we boot
- delete mpfs_rtc_init & set prescale in probe
- use dev_pm_set_wake_irq rather than writing suspend/resume functions

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>